requirements.txt
Appium-Python-Client==4.0.1
selenium==4.23.1

